Make ffmpeg optional feature
A lot of users only need this crate for asserting images, not for asserting video. Those users shouldn't require ffmpeg.
