+++ Cargo.toml
@@ -2,7 +2,7 @@
 name = "twenty-twenty"
 description = "Visual regression testing of H264 frames and images."
 repository = "https://github.com/kittycad/twenty-twenty"
-version = "0.6.1"
+version = "0.7.0"
 edition = "2021"
 license = "MIT"
 
@@ -10,7 +10,11 @@ license = "MIT"
 
 [dependencies]
 anyhow = { version = "1.0.71", features = ["backtrace"] }
-ffmpeg-next = { version = "6.0.0", default-features = false, features = ["codec", "format", "software-scaling"] }
+ffmpeg-next = { version = "6.0.0", default-features = false, features = ["codec", "format", "software-scaling"], optional = true }
 image = "0.24.6"
 image-compare = "0.3.0"
 uuid = { version = "1.4.0", features = ["v4"] }
+
+[features]
+default = []
+h264 = ["dep:ffmpeg-next"]
